Cleanup Static vs Interactive section
