+++ packages/react-openapi/src/InteractiveSection.tsx
@@ -1,71 +1,15 @@
 'use client';
 
 import clsx from 'clsx';
-import { type ComponentPropsWithoutRef, forwardRef, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 import { mergeProps, useButton, useDisclosure, useFocusRing } from 'react-aria';
 import { useDisclosureState } from 'react-stately';
+import { Section, SectionBody, SectionHeader, SectionHeaderContent } from './StaticSection';
 
 interface InteractiveSectionTab {
     key: string;
     label: string;
     body: React.ReactNode;
-}
-
-export function Section(props: ComponentPropsWithoutRef<'div'>) {
-    return <div {...props} className={clsx('openapi-section', props.className)} />;
-}
-
-export function SectionHeader(props: ComponentPropsWithoutRef<'div'>) {
-    return (
-        <div
-            {...props}
-            className={clsx(
-                'openapi-section-header',
-                props.className && `${props.className}-header`
-            )}
-        />
-    );
-}
-
-export function SectionHeaderContent(props: ComponentPropsWithoutRef<'div'>) {
-    return (
-        <div
-            {...props}
-            className={clsx(
-                'openapi-section-header-content',
-                props.className && `${props.className}-header-content`
-            )}
-        />
-    );
-}
-
-export const SectionBody = forwardRef(function SectionBody(
-    props: ComponentPropsWithoutRef<'div'>,
-    ref: React.ForwardedRef<HTMLDivElement>
-) {
-    return (
-        <div
-            ref={ref}
-            {...props}
-            className={clsx('openapi-section-body', props.className && `${props.className}-body`)}
-        />
-    );
-});
-
-export function StaticSection(props: {
-    className: string;
-    header: React.ReactNode;
-    children: React.ReactNode;
-}) {
-    const { className, header, children } = props;
-    return (
-        <Section className={className}>
-            <SectionHeader className={className}>
-                <SectionHeaderContent className={className}>{header}</SectionHeaderContent>
-            </SectionHeader>
-            <SectionBody className={className}>{children}</SectionBody>
-        </Section>
-    );
 }
 
 /**
